requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

langgraph==0.6.7

openai==1.55.3

sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.1

pinecone==7.3.0

tavily-python==0.3.3
requests

pandas==2.1.4
numpy==1.26.2

python-dotenv==1.0.0

structlog==23.2.0

pytest==7.4.3
black==23.12.0
flake8

cryptography

pydantic
python-dateutil==2.8.2
pillow==10.1.0

scikit-learn
mysql-connector-python
pydantic[email]
aiohttp
tavily-python
python-multipart
faster-whisper
bs4
pyjwt
lxml
